Argument setup, CMD, and OP code implementation
Still super buggy and half-baked, but this stuff works better now. Fixed the bug with file streams not being passed by reference. Added some functions to determine which type of instruction is being dealt with (or if it is a label), to extract the register identifier from the assembly, find the CMD field for a data-processing instruction, andto snatch the first group of characters in a string before hitting whitespace. These were all added to the header. In addition, I think I cleaned up the logic for casting the type of instruction to an enum for visibility, and I fixed the variable names associated with that. The switch statement at the core of the main loop is slightly more fleshed out, able to do different things depending on the presnece of immediates or links, and the main loop is now a for loop to keep track of instruction addresses, not a do-while.