Compiles because files moved
Previously, the Makefile was in the parent directory containing the .xcodeproj and subfolder with all the code, but that has since been rectified. I fixed the main program to also include the header, not the whole implementation file. The program compiles both with the current makefile AND under Xcode's propreitary build tools, so it cannot be too terribly busted. I have moved the local copy to a GitHub clone becasue the other one would not fetch updates or display all the files correctly after doing some light edits on my PC. The executable for my OS (macOS 14.0 Intel x86) is also uploaded for reference. I will omit the object files because they are not that interesting or useful to humans.