Fleshed Out Instruction Classes
The main upgrade of this commit is that I have successfully compiled the new object-oriented structure of the program in both Xcode and with a makefile. For the most part, all of the attributes and member functions of each class should be finished, so now I can work on cleaning up error handling and unused code in 'finalFunctions' and on finishing the logic in 'main', which will mostly involve reading information to the appropriate variables form the string stream. I used some clever tricks in this update, notably by telling the extraction operator to check every time for whether it is reading a hexedecimal or integer value when storing to an integral data type, which, combined with the capability of bitset, means I barely have to do ANY base conversion. The main obstacle to making all my code compile was my previous approach of trying to use generic programming to concatenate bitsets of unknown sizes, which caused the linker to pitch a complete and total fit. I have replaced all that code with virtual functions that make the parent do as much of the common work as possible and copy the information from different attributes into the binary encoding of the full instruction. I have made extensive use of my scrubString(string& info) function that removes any 'R', '#', or ',' characters from the string and converts it to an integer after checking what base it is in (decimal of hex). Updating the makefile was probably the most tedious part, as Xcode has no syntax highlighting for anything outside of its propreitary Apple walled garden, which enabled me to spend 15-20 minutes hunting down spelling errors and bits I forgot to change after copy-pasting. The project compiles, but throws errors whenever I try to open an input file, which is what I told it to do but means I translated my intent incorrectly or missed something. Still, the binary is built for Intel x86 macOS >= 14.0 systems (still named "assembler") and can be built for others as well. I am beginning to worry about the size and scope of this project brought about by my extreme tryhard tendencies, as Xcode struggles to render all the changes when I scroll down my commit list on my MacBook Pro from 2020 with a whole 16 GB of RAM. This has gotten out of hand, but it'll all be over regardless when this is due Tuesday...